pyproject.toml
[project]
name = "comicbox-pdffile"
version = "0.2.5"
authors = [{ name = "AJ Slater", email = "aj@slater.net" }]
license = { text = "GPL-3.0-only" }
requires-python = "<4.0,>=3.9"
dependencies = [
  "filetype<2.0.0,>=1.2.0",
  "pymupdf<2.0.0,>=1.24.0",
  "python-dateutil>=2.9.0.post0",
  "typing-extensions>=4.13.2",
]
description = "A ZipFile like API for PyMuPDF"
readme = "README.md"
keywords = []
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
]

[project.urls]
News = "https://github.com/ajslater/comicbox-pdffile"
"Report Issues" = "https://github.com/ajslater/comicbox-pdffile/issues"
homepage = "https://github.com/ajslater/comicbox-pdffile"
repository = "https://github.com/ajslater/comicbox-pdffile"
documentation = "https://github.com/ajslater/comicbox-pdffile"

[dependency-groups]
dev = [
  "basedpyright~=1.29.1",
  "codespell<3.0.0,>=2.1.0",
  "coverage[toml]<8.0,>=7.0",
  "hatchling<2.0.0,>=1.27.0",
  "pytest>=8.3.5",
  "pytest-cov<7.0.0,>=6.0.0",
  "pytest-gitignore<2.0,>=1.3",
  "radon[toml]<7.0.0,>=6.0.1",
  "ruff<1.0.0,>=0.9.2",
  "toml-cli<1.0.0,>=0.7.0",
  "types-python-dateutil<3.0.0,>=2.8.19",
  "vulture<3.0,>=2.3",
  "wheel<1.0.0,>=0.45.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pdffile"]

[tool.hatch.build.targets.sdist]
include = ["pdffile", "tests", "NEWS.md"]
exclude = ["*~"]

[tool.basedpyright]
exclude = [
  "**/.*",
  "**/__pycache__",
  "**/node_modules",
  "dist",
  "test-results",
  "typings",
  "vulture_whitelist.py",
]
pythonVersion = "3.10"
failOnWarnings = false
reportAny = false
reportExplicitAny = false
reportImportCycles = true
reportIncompatibleUnannotatedOverride = true
reportMissingParameterType = false
reportMissingTypeArgument = false
reportPrivateUsage = false # ruff does this
reportUnannotatedClassAttribute = false
reportUnknownArgumentType = false
reportUnknownMemberType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportUnusedCallResult = false
reportUnusedParameter = false # ruff does this 
reportMissingTypeStubs = false

[tool.codespell]
skip = ".*,*~,./dist,./node_modules,./package.json,./package-lock.json,./uv.lock,./test-results"
builtin = 'clear,rare,code'
check-hidden = true
# Remove 'coverd' when this issue is resolved
# https://github.com/codespell-project/codespell/issues/1212
ignore-words-list = 'thead,ro,jupyter,coverd,searched,ws,falsy'

[tool.coverage.html]
directory = "test-results/coverage"

[tool.coverage.run]
source = ["."]
branch = true
concurrency = ["multiprocessing"]
omit = [
  "*/.*",
  "*__pycache__*",
  "dist/*",
  "node_modules/*",
  "test-results/*",
  "typings/*",
]

[tool.pytest.ini_options]
addopts = """
    --junit-xml=test-results/pytest/results.xml
    -ra
    --strict-config
    --strict-markers
    --cov
    --cov-append
    --cov-report=html
    --cov-report=term
"""
junit_family = "xunit2"
testpaths = "tests"
pythonpath = ["."]

[tool.radon]
exclude = "*~,.*,.*/*,__pycache__/*,dist/*,node_modules/*,test-results/*,typings/*"

[tool.ruff]
extend-exclude = ["**/.*", "**/__pycache__", "dist", "test-results", "typings"]
target-version = "py39"

[tool.ruff.lint]
extend-ignore = [
  "BLE001",  # catch bare exception
  "COM812",  # missing trailing comma
  "COM819",  # prohibit trailing comma
  "D203",    # blank line before class docstring
  "D206",    # formatted
  "D212",    # first line multi line docstring.
  "E111",    # formatted
  "E114",    # formatted
  "E117",    # over-indented
  "E501",    # 88 char line
  "ISC001",  # formatted
  "PERF203", # try block in loop
  "S101",    # assert
  "W191",    # formatted
]
extend-select = [
  "A",
  # "AIR",
  # "ANN",
  "ARG",
  "ASYNC",
  "B",
  "B9",
  "BLE",
  "C",
  "C4",
  "C90",
  "COM",
  # "CPY",
  "D",
  "DJ",
  # "DOC",
  "DTZ",
  "E",
  "EM",
  "ERA",
  "EXE",
  "F",
  "FA",
  # "FAST",
  "FBT",
  "FIX",
  "FLY",
  "FURB",
  # "G",
  "I",
  "ICN",
  "INP",
  "INT",
  "ISC",
  "LOG",
  # "PD",
  "PERF",
  "PGH",
  "PIE",
  "PL",
  "PT",
  "PTH",
  "PYI",
  # "Q", Formatted
  "N",
  # NPY,
  "RET",
  "RSE",
  "RUF",
  "S",
  "SIM",
  "SLF",
  "SLOT",
  "T10",
  "T20",
  "TCH",
  # "TD",
  "TID",
  "TRY",
  "UP",
  "W",
  "YTT",
]
external = ["V101"]
task-tags = ["TODO", "FIXME", "XXX", "http", "HACK"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["SLF001", "T201", "T203"]

[tool.ruff.lint.pycodestyle]
ignore-overlong-task-comments = true

[tool.vulture]
exclude = [
  "*/.*",
  "*/__pycache__*",
  "*/node_modules*",
  "dist/",
  "test-results/",
  "typings/",
]
min_confidence = 61
sort_by_size = true
